add php

+++ admin/pages_loans.php
@@ -28,12 +28,12 @@ $admin_id = $_SESSION['admin_id'];
         <div class="container-fluid">
           <div class="row mb-2">
             <div class="col-sm-6">
-              <h1>iBanking Loans</h1>
+              <h1>ACLEDA BANK Plc. Loans</h1>
             </div>
             <div class="col-sm-6">
               <ol class="breadcrumb float-sm-right">
                 <li class="breadcrumb-item"><a href="pages_dashboard.php">Dashboard</a></li>
-                <li class="breadcrumb-item"><a href="pages_manage_clients.php">iBank Loans</a></li>
+                <li class="breadcrumb-item"><a href="pages_manage_clients.php">ACLEDA BANK Plc. Loans</a></li>
                 <li class="breadcrumb-item active">Give Loans</li>
               </ol>
             </div>
@@ -65,7 +65,7 @@ $admin_id = $_SESSION['admin_id'];
                   </thead>
                   <tbody>
                     <?php
-                    //fetch all iBank clients
+                    //fetch all ACLEDA BANK Plc. clients
                     $ret = "SELECT * FROM  iB_clients ORDER BY RAND() ";
                     $stmt = $mysqli->prepare($ret);
                     $stmt->execute(); //ok
@@ -90,7 +90,7 @@ $admin_id = $_SESSION['admin_id'];
                             Manage
                           </a>
 
-                          <a class="btn btn-danger btn-sm" href="pages_manage_clients.php?deleteClient=<?php echo $row->client_id; ?>">
+                          <a class="btn btn-danger btn-sm" href="pages_manage_clients.php?deleteClient=<?php echo $row->client_id; ?>" onclick="event.preventDefault(); confirmDelete(this.href);">
                             <i class="fas fa-trash"></i>
                             <!-- <i class="fas fa-user"></i> -->
                             Delete
